requirements.txt
flask
flask-socketio
pyopenssl
pyngrok
simple-websocket
python-dotenv